requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
Pillow>=10.0.0  # для работы с изображениями
djangorestframework-simplejwt>=5.2.2
django-cors-headers>=4.1.0
psycopg2-binary>=2.9.6  # для работы с PostgreSQL
python-dotenv>=1.0.0  # для работы с переменными окружения 